requirements.txt
asgiref==3.3.1
Django==3.1.6
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
django-simple-captcha~=0.5.14
keyring~=23.0.1
requests~=2.26.0